AI_Presenter/requirements.txt
ollama==0.3.1
azure-cognitiveservices-speech==1.38.0
deepgram-sdk==3.3.0
groq==0.9.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.28
langchain-groq==0.1.5
langchain-ollama==0.1.1
langchain-text-splitters==0.2.2
langsmith==0.1.98
python-pptx==1.0.1
pyttsx3==2.90
pydub==0.25.1
python-multipart==0.0.9
pyaudio==0.2.14
python-dotenv
img2pdf 
Pillow
fastapi
uvicorn
openai
PyPDF2>=3.0.0
loguru
